+++ camo/estimate/effects.py
@@ -4,6 +4,8 @@ from typing import Any, Tuple
 import numpy as np
 import pandas as pd
 
+from tqdm import trange
+
 from . import methods
 from ..utils import _to_categorical, _try_get
 
@@ -17,10 +19,13 @@ def average_causal_effect(
     Z: str = None,
     method: str = "g_formula",
     estimator: str = "ols",
-    targets: Tuple[Any] = (0, 1)
+    targets: Tuple[Any] = (0, 1),
+    bootstrap: int = None,
+    alpha: float = 0.05
 ) -> float:
     # Try get value from methods
-    method = _try_get(method, methods)
+    if isinstance(method, str):
+        method = _try_get(method, methods)
 
     # Check if data is numeric
     is_numeric = np.vectorize(lambda x: not np.issubdtype(x, np.number))
@@ -34,7 +39,31 @@ def average_causal_effect(
     effect_0, effect_1 = method(data, X, Y, Z, estimator)
 
     # ACE = E_Z[ E[Y|do(X=1),Z] - E[Y|do(X=0),Z] ]
-    return np.mean(effect_1 - effect_0)
+    ace = np.mean(effect_1 - effect_0)
+
+    # If bootstrap sample size is specified compute confidence interval
+    if bootstrap:
+        # Preallocate samples vector
+        samples = np.empty((bootstrap, ))
+        # For each bootstrap sample
+        for i in trange(bootstrap):
+            # Sample from dataset with replacement
+            sample = data.sample(len(data), replace=True)
+            sample.reset_index(drop=True, inplace=True)
+            # Call ACE recursively
+            samples[i] = average_causal_effect(
+                sample,
+                X, Y, Z,
+                method,
+                estimator,
+                targets
+            )
+        # Compute upper and lower bounds over samples
+        quantiles = np.quantile(samples, q=[alpha/2, 1-alpha/2])
+        # Return ACE with confidence bounds
+        return ace, *quantiles
+    
+    return ace
 
 
 def total_effect() -> float: